Updated RESET and STATUS pin usage
Using the same pin for RESET and STATE control.
This pin is normally configured as INPUT to read modem state.
To reset the modem we should force a LOW level on this pin and the switch
it to be an OUTPUT.
NOTE: we should strictly avoid to force HIGH levels on RESET/STATUS pin

git-svn-id: file:///media/hdbkp/svn/c3/trunk/cbox@63 c892b811-77a8-453b-ad2f-bbcecf767e94