Added new Linux Event based Signal implementation
This patch add a Signal implementation based on the standard Linux Event
interface. Hardware generated interrupts are associated to GPIO pins,
and thus are handled has a keyboard.

This solution should grant higher portability among different platforms.
Pheraps we should provide a Linux driver that maps GPIO to generic
events without using the keyboard interface that can be effectively used
for this purpose.